ConfigDialog: Add "iMotions Events API" as option for eye-tracker
Note that the order of eye-trackers had to change, because Tobii
is more restrictive: tobii_research module is harder to install
because it requires specific Python version (Python 3.10 or 3.8),
and AvailibilityChecker.checkEyeTracker() actually checks if there
are any Tobii trackers available.

Optional values must go last, at least in current implementation.
The "iMotions Events API" eye-tracker is currently added
unconditionally; there is no check that this API is available (yet).

Not tested.