WIP (stimuliWatcher.start in background thread)
https://plugins.jetbrains.com/docs/intellij/background-processes.html

It does not block the main thread even in the presence of the long
Thread.sleep().

.
However, with the old implementation of emulate_imotions.py script
that was used in place of actual iMotions Lab aplication, it did not
correctly process data from this iMotions Lab emulator, closing the
client after stimuli start.

This was caused by the server not handling the check for server
liveness, where the client aborted connection after reading one byte,
correctly.  The server stopped, and thus the stimuli watcher did not
work as intended.

Notifications (for old emulate_imotions.py):

  Auto tracking enabled
  stimuliWatcher: Stimulus presentation started
  stimuliWatcher: iMotions Lab server disconnected
  slept with Thread.sleep()
  after Thread.sleep()